last dan first di paginator

+++ application/modules/admin/controllers/ArticleController.php
@@ -5,6 +5,7 @@ class Admin_ArticleController extends Zend_Controller_Action
 //  messages
   const MSG_SELECTED_ARTICLES_DELETED = 'success|Artikel yang dipilih berhasil dihapus.';
   const MSG_ARTICLE_CREATED = 'success|Artikel berhasil dibuat.';
+  const MSG_ARTICLE_ARCHIVED = 'success|Artikel dipindahkan ke arsip.';
   const MSG_ARTICLE_EDITED = 'success|Artikel berhasil disunting.';
   const MSG_ARTICLE_DELETED = 'success|Artikel berhasil dihapus.';
   const MSG_ARTICLE_RESTORED = 'success|Artikel berhasil dikembalikan.';
@@ -152,7 +153,9 @@ class Admin_ArticleController extends Zend_Controller_Action
                 'title' => $this->form->title->getValue(),
                 'content' => $this->form->content->getValue(),
                 'tags' => $this->form->tags->getValue(),
-                'status' => $status
+                'status' => $status,
+                'updated_by' => Zend_Auth::getInstance()->getIdentity()->username,
+                'updated_on' => Date('Y-m-d H:i:s')
             ))->save();
 
             $this->_helper->flashMessenger->addMessage
@@ -174,10 +177,17 @@ class Admin_ArticleController extends Zend_Controller_Action
     if (null !== $id) {
       $article = $this->article->find($id)->current();
       if (null !== $article) {
-        $this->tag->save('', $article->tags);
-        $article->delete();
-        $this->_helper->flashMessenger->addMessage
-                (self::MSG_ARTICLE_DELETED);
+        if (Admin_Model_Status::ARCHIVED === $article->status) {
+          $this->tag->save('', $article->tags);
+          $article->delete();
+          $this->_helper->flashMessenger->addMessage
+                  (self::MSG_ARTICLE_DELETED);
+        } else {
+          $article->status = Admin_Model_Status::ARCHIVED;
+          $article->save();
+          $this->_helper->flashMessenger->addMessage
+                  (self::MSG_ARTICLE_ARCHIVED);
+        }
       }
     }
     $this->_helper->redirector('index');
